Toggled main network structures to Array struct

+++ src/network/utils/cost.cpp
@@ -1,6 +1,6 @@
-double countCost(double *output, double *expected, int columnsAmount)
+float countCost(float *output, float *expected, int columnsAmount)
 {
-  double cost = 0.0;
+  float cost = 0.0;
   for (int i = 0; i < columnsAmount; i++)
     cost += (expected[i] - output[i]) * (expected[i] - output[i]);
   return cost;
